fix: 유튜브 목록 조회 시 스크랩수 및 조회수 정렬 오류 수정

+++ src/main/java/com/recipe/app/src/recipe/domain/youtube/YoutubeRecipe.java
@@ -42,10 +42,10 @@ public class YoutubeRecipe extends BaseEntity {
     private String youtubeId;
 
     @Column(name = "scrapCnt", nullable = false)
-    private int scrapCnt;
+    private long scrapCnt;
 
     @Column(name = "viewCnt", nullable = false)
-    private int viewCnt;
+    private long viewCnt;
 
     @Builder
     public YoutubeRecipe(Long youtubeRecipeId, String title, String description, String thumbnailImgUrl, LocalDate postDate, String channelName, String youtubeId) {
